pyproject.toml
[project]
name = "play-glm-orc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glmocr>=0.1.4",
    "ollama>=0.6.1",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
]
